requirements.txt
beautifulsoup4==4.12.2
# blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
# charset-normalizer==3.2.0
# click==8.1.6
# colorama==0.4.6
confection==0.1.0
# cssselect==1.2.0
cymem==2.0.7
# dnspython==2.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.1
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pathy==0.10.2
phonenumbers==8.13.17
Pillow==10.0.0
pydantic==1.10.12
pymongo==4.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
preshed==3.0.8
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.10
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.6